Set up code to for a custom module
