+++ custom_modules/custom_queries/custom_query_library.py
@@ -1,0 +1,20 @@
+from dataclasses import dataclass
+from typing import Dict, Optional
+
+from string import Template
+
+from promg import Query
+
+
+class CustomCypherQueryLibrary:
+    @staticmethod
+    def get_example_query(labels: str):
+        query_str = '''
+            MATCH (n:$labels)
+            RETURN count(n)
+        '''
+
+        return Query(query_str=query_str,
+                     template_string_parameters={
+                        "labels": labels
+        })
